+++ CHANGELOG.md
@@ -22,7 +22,6 @@
 - Added support for `Session.directory` and `Session.read.directory` to retrieve the list of all files on a stage with metadata.
 - Added support for `DataFrameReader.jdbc`(PrPr) that allows ingesting external data source with jdbc driver.
 - Added support for `FileOperation.copy_files` to copy files from a source location to an output stage.
-
 - Added support for the following scalar functions in `functions.py`:
   - `all_user_names`
   - `current_account_name`
@@ -33,6 +32,7 @@
   - `current_organization_name`
   - `current_organization_user`
   - `current_transaction`
+- Added support for `Session.begin_transaction`, `Session.commit` and `Session.rollback`.
 
 #### Bug Fixes